refactor: improve docs, DX, and fork-friendliness
- Centralize site title/bio/description in src/data/site.ts; all
  layouts and components read from it instead of hardcoding
- Derive canonical/OG URLs from Astro.site instead of hardcoding domain
- Replace static robots.txt with dynamic Astro API route
- Reorganize src/themes/: definitions/ subfolder, index.ts → registry.ts
- Move unit test to tests/unit/, rename to match function under test
- Slim README, deduplicate with CONTRIBUTING.md, fix npm → pnpm
- Add FORKING.md guide for personalizing a fork
- Fix .nvmrc (was 11.15.0, now 22 to match CI)
- Fix ESLint astro parser config for TypeScript frontmatter
- Add checklist to PR template
