+++ .github/pull_request_template.md
@@ -1,3 +1,7 @@
 ## Description
 
+<!-- What does this PR do and why? -->
+
 ## Testing
+
+<!-- How was this tested? -->
