+++ .nvmrc
@@ -1,1 +1,1 @@
-11.15.0
+22
