+++ eslint.config.js
@@ -8,6 +8,14 @@ export default [
   js.configs.recommended,
   ...tseslint.configs.recommended,
   ...astro.configs.recommended,
+  {
+    files: ['**/*.astro'],
+    languageOptions: {
+      parserOptions: {
+        parser: tseslint.parser,
+      },
+    },
+  },
   unicorn.configs['flat/recommended'],
   {
     languageOptions: {
