+++ src/components/ThemeSwitcher.astro
@@ -16,7 +16,7 @@
 </div>
 
 <script>
-  import { themes, applyTheme, getStoredThemeId } from '../themes/index';
+  import { themes, applyTheme, getStoredThemeId } from '../themes/registry';
 
   const switcher = document.querySelector('.theme-switcher');
   const btn = document.querySelector('.theme-btn');
