+++ src/data/site.ts
@@ -1,5 +1,7 @@
 export const site = {
   title: 'colin tinney',
+  bio: 'Software engineer. Currently working at <a href="https://squareup.com">Block (formerly known as Square)</a>.',
+  description: 'A personal website',
   contact: {
     email: 'colintinney@gmail.com',
     displayedEmail: 'colintinney [at] gmail.com',
