+++ src/pages/404.astro
@@ -1,6 +1,6 @@
 ---
 import PageLayout from '../layouts/PageLayout.astro';
-import { themes } from '../themes/index';
+import { themes } from '../themes/registry';
 
 const themeEntries = Object.entries(themes);
 ---
